# swagger.yaml
...
openapi: 3.0.3
info:
  title: Clinic-AI API
  version: 1.0.0
  description: "AI-powered clinical intake system. This spec is being filled module-by-module."
servers:
  - url: http://localhost:8000
paths:
...
  /audio/:
    get:
      tags: [audio]
      summary: List audio files
      operationId: list_audio_files
      parameters:
        - name: patient_id
          in: query
          required: false
          schema: {type: string}
        - name: visit_id
          in: query
          required: false
          schema: {type: string}
        - name: audio_type
          in: query
          required: false
          schema: {type: string, description: 'adhoc | visit'}
        - name: limit
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
        - name: offset
          in: query
          required: false
          schema: {type: integer, minimum: 0, default: 0}
      responses:
        '200':
          description: Audio files listed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AudioListResponse'
        '500': {description: Failed to list audio files, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /audio/dialogue:
    get:
      tags: [audio]
      summary: List structured dialogues for audio files
      operationId: list_audio_dialogues
      parameters:
        - name: patient_id
          in: query
          required: false
          schema: {type: string}
        - name: visit_id
          in: query
          required: false
          schema: {type: string}
        - name: audio_type
          in: query
          required: false
          schema: {type: string}
        - name: start_date
          in: query
          required: false
          schema: {type: string, format: date-time}
        - name: end_date
          in: query
          required: false
          schema: {type: string, format: date-time}
        - name: limit
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
        - name: offset
          in: query
          required: false
          schema: {type: integer, minimum: 0, default: 0}
      responses:
        '200':
          description: Audio dialogues listed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AudioDialogueListResponse'
        '500': {description: Failed to list audio dialogues, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /audio/{audio_id}:
    get:
      tags: [audio]
      summary: Get audio file metadata
      operationId: get_audio_metadata
      parameters:
        - name: audio_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Audio metadata retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AudioFileResponse'
        '404': {description: Audio not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Failed to get audio file, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    delete:
      tags: [audio]
      summary: Delete audio file
      operationId: delete_audio_file
      parameters:
        - name: audio_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '204': {description: Deleted}
        '404': {description: Audio not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Failed to delete audio file, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /audio/{audio_id}/download:
    get:
      tags: [audio]
      summary: Download audio file
      operationId: download_audio_file
      parameters:
        - name: audio_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200': {description: Binary stream}
        '404': {description: Audio not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Failed to download audio file, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /audio/{audio_id}/stream:
    get:
      tags: [audio]
      summary: Stream audio file
      operationId: stream_audio_file
      parameters:
        - name: audio_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200': {description: Binary stream}
        '404': {description: Audio not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Failed to stream audio file, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /audio/stats/summary:
    get:
      tags: [audio]
      summary: Get audio storage statistics
      operationId: get_audio_stats
      responses:
        '200': {description: Stats returned, content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse_Dict'}}}}
        '500': {description: Failed to get audio statistics, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
components:
  schemas:
    ErrorResponse:
      type: object
      required: [success, error, message, timestamp, request_id]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: INTERNAL_SERVER_ERROR
        message:
          type: string
          example: An unexpected internal server error occurred.
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
...
    ApiResponse_Dict:
      type: object
      required: [success, message, timestamp, request_id, data]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: OK
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
        data:
          type: object
          additionalProperties: true
...
    AudioFileResponse:
      type: object
      required: [audio_id, filename, content_type, file_size, audio_type, created_at, updated_at]
      properties:
        audio_id: {type: string}
        filename: {type: string}
        content_type: {type: string}
        file_size: {type: integer}
        duration_seconds: {type: number, nullable: true}
        patient_id: {type: string, nullable: true}
        visit_id: {type: string, nullable: true}
        adhoc_id: {type: string, nullable: true}
        audio_type: {type: string}
        created_at: {type: string}
        updated_at: {type: string}
    AudioListResponse:
      type: object
      required: [files, total_count, limit, offset]
      properties:
        files:
          type: array
          items: {$ref: '#/components/schemas/AudioFileResponse'}
        total_count: {type: integer}
        limit: {type: integer}
        offset: {type: integer}
    AudioDialogueResponse:
      type: object
      required: [audio_id, filename, audio_type, created_at]
      properties:
        audio_id: {type: string}
        filename: {type: string}
        duration_seconds: {type: number, nullable: true}
        patient_id: {type: string, nullable: true}
        visit_id: {type: string, nullable: true}
        adhoc_id: {type: string, nullable: true}
        audio_type: {type: string}
        created_at: {type: string}
        structured_dialogue:
          type: array
          items: {type: object, additionalProperties: true}
          nullable: true
    AudioDialogueListResponse:
      type: object
      required: [dialogues, total_count, limit, offset]
      properties:
        dialogues:
          type: array
          items: {$ref: '#/components/schemas/AudioDialogueResponse'}
        total_count: {type: integer}
        limit: {type: integer}
        offset: {type: integer}
    ApiResponse_AudioListResponse:
      type: object
      required: [success, message, timestamp, request_id]
      properties:
        success: {type: boolean, example: true}
        message: {type: string, example: 'Audio files listed'}
        timestamp: {type: string, format: date-time}
        request_id: {type: string, format: uuid}
        data:
          $ref: '#/components/schemas/AudioListResponse'
    ApiResponse_AudioDialogueListResponse:
      type: object
      required: [success, message, timestamp, request_id]
      properties:
        success: {type: boolean, example: true}
        message: {type: string, example: 'Audio dialogues listed'}
        timestamp: {type: string, format: date-time}
        request_id: {type: string, format: uuid}
        data:
          $ref: '#/components/schemas/AudioDialogueListResponse'
    ApiResponse_AudioFileResponse:
      type: object
      required: [success, message, timestamp, request_id]
      properties:
        success: {type: boolean, example: true}
        message: {type: string, example: 'Audio metadata retrieved'}
        timestamp: {type: string, format: date-time}
        request_id: {type: string, format: uuid}
        data:
          $ref: '#/components/schemas/AudioFileResponse'
...